+++ backend/canisters/user/impl/src/updates/create_group.rs
@@ -13,7 +13,7 @@ use utils::text_validation::{
     validate_description, validate_group_name, validate_rules, NameValidationError, RulesValidationError,
 };
 
-#[update(guard = "caller_is_owner", candid = true, msgpack = true)]
+#[update(guard = "caller_is_owner", msgpack = true)]
 #[trace]
 async fn create_group(mut args: Args) -> Response {
     run_regular_jobs();